fix: refactoring for EB
